+++ frontend/s_a_m_e/lib/account/login.dart
@@ -2,11 +2,11 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:fluttertoast/fluttertoast.dart';
-import 'package:s_a_m_e/admin.dart';
+import 'package:s_a_m_e/admin/admin_home.dart';
 import 'package:s_a_m_e/colors.dart';
-import 'package:s_a_m_e/signup.dart';
-import 'package:s_a_m_e/user_home.dart';
-import 'package:s_a_m_e/firebase_service.dart';
+import 'package:s_a_m_e/account/signup.dart';
+import 'package:s_a_m_e/user/user_home.dart';
+import 'package:s_a_m_e/firebase/firebase_service.dart';
 
 class Login extends StatefulWidget {
   const Login({Key? key}) : super(key: key);
