+++ frontend/s_a_m_e/lib/account/view_profile.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:fluttertoast/fluttertoast.dart';
 import 'package:s_a_m_e/colors.dart';
-import 'package:s_a_m_e/firebase_service.dart';
+import 'package:s_a_m_e/firebase/firebase_service.dart';
 
 class ProfilePage extends StatelessWidget {
