+++ frontend/s_a_m_e/lib/admin/add_category.dart
@@ -1,26 +1,26 @@
 import 'package:flutter/material.dart';
 import 'package:multi_select_flutter/multi_select_flutter.dart';
-import 'package:s_a_m_e/symptomlist.dart';
+import 'package:s_a_m_e/user/categorieslist.dart';
+import 'package:s_a_m_e/user/symptomlist.dart';
 import 'package:s_a_m_e/colors.dart';
-import 'package:s_a_m_e/firebase_service.dart';
-import 'package:s_a_m_e/profilepicture.dart';
+import 'package:s_a_m_e/firebase/firebase_service.dart';
+import 'package:s_a_m_e/account/profilepicture.dart';
 
-class SymptomCreationPage extends StatefulWidget {
-  const SymptomCreationPage({super.key});
+class CategoryCreationPage extends StatefulWidget {
+  const CategoryCreationPage({Key? key}) : super(key: key);
 
   @override
-  _SymptomCreationPageState createState() => _SymptomCreationPageState();
+  _CategoryCreationPageState createState() => _CategoryCreationPageState();
 }
 
-class _SymptomCreationPageState extends State<SymptomCreationPage> {
-  //final _apiService = ApiService();
-  final _symptomNameController = TextEditingController();
+class _CategoryCreationPageState extends State<CategoryCreationPage> {
+  final TextEditingController _categoryNameController = TextEditingController();
   final FirebaseService _firebaseService = FirebaseService();
-  List<Category> _selectedComplaints = [];
+  List<String> _selectedSymptoms = [];
 
   @override
   void dispose() {
-    _symptomNameController.dispose();
+    _categoryNameController.dispose();
     super.dispose();
   }
 
@@ -29,21 +29,19 @@ class _SymptomCreationPageState extends State<SymptomCreationPage> {
     return Scaffold(
       appBar: AppBar(
         title: const Text('S.A.M.E'),
-        // ignore: prefer_const_constructors, prefer_const_literals_to_create_immutables
-        // actions: [ProfilePicturePage()]
       ),
       body: Padding(
         padding: const EdgeInsets.all(15.0),
         child: Column(
           children: [
             const SizedBox(height: 20),
-            const Text('Add Symptom', style: TextStyle(fontWeight: FontWeight.bold, fontSize: 28.0)),
+            const Text('Add Category', style: TextStyle(fontWeight: FontWeight.bold, fontSize: 28.0)),
             const SizedBox(height: 40),
             TextField(
-              controller: _symptomNameController,
+              controller: _categoryNameController,
               decoration: const InputDecoration(
                 contentPadding: EdgeInsets.all(20.0),
-                labelText: 'Symptom Name',
+                labelText: 'Category Name',
                 labelStyle: TextStyle(color: navy),
                 filled: true,
                 fillColor: boxinsides,
@@ -58,59 +56,61 @@ class _SymptomCreationPageState extends State<SymptomCreationPage> {
               ),
             ),
             const SizedBox(height: 30),
-            FutureBuilder<List<Category>>(
-              future: _firebaseService.getAllCategories(),
-              builder: (context, categoriesSnapshot) {
-                if (categoriesSnapshot.connectionState == ConnectionState.waiting) {
+            FutureBuilder<List<String>>(
+              future: _firebaseService.getAllSymptoms(),
+              builder: (context, symptomsSnapshot) {
+                if (symptomsSnapshot.connectionState == ConnectionState.waiting) {
                   return const CircularProgressIndicator();
-                } else if (categoriesSnapshot.hasError) {
-                  return Text('Error: ${categoriesSnapshot.error}');
+                } else if (symptomsSnapshot.hasError) {
+                  return Text('Error: ${symptomsSnapshot.error}');
                 } else {
-                  List<Category>? categories = categoriesSnapshot.data;
+                  List<String>? symptoms = symptomsSnapshot.data;
 
-                  if (categories != null && categories.isNotEmpty) {
-                    return MultiSelectDialogField<Category>(
+                  if (symptoms != null && symptoms.isNotEmpty) {
+                    return MultiSelectDialogField<String>(
                       backgroundColor: background,
                       cancelText: const Text('CANCEL', style: TextStyle(fontWeight: FontWeight.bold, color: navy)),
                       confirmText: const Text('SELECT', style: TextStyle(fontWeight: FontWeight.bold, color: navy)),
                       unselectedColor: navy,
                       selectedColor: navy,
-                      items: categories
-                          .map((complaint) => MultiSelectItem<Category>(
-                              complaint, complaint.name))
+                      items: symptoms
+                          .map((symptom) => MultiSelectItem<String>(
+                              symptom, symptom))
                           .toList(),
-                      title: const Text("Categories"),
+                      title: const Text("Symptoms"),
                       onConfirm: (values) {
-                        _selectedComplaints = values;
+                        setState(() {
+                          _selectedSymptoms = values;
+                        });
                       },
                     );
                   } else {
-                    return const Text('No categories available');
+                    return const SizedBox(); // Return an empty SizedBox if no symptoms available
                   }
                 }
               },
             ),
+
             const SizedBox(height: 20),
             ElevatedButton(
               style: const ButtonStyle(
-                  foregroundColor: MaterialStatePropertyAll<Color>(white),
-                  backgroundColor: MaterialStatePropertyAll<Color>(navy),
-                ),
+                foregroundColor: MaterialStatePropertyAll<Color>(white),
+                backgroundColor: MaterialStatePropertyAll<Color>(navy),
+              ),
               onPressed: () async {
-                if (_symptomNameController.text.isNotEmpty &&
-                    _selectedComplaints.isNotEmpty) {
-                  final response = await _firebaseService.addSymptom(
-                    _symptomNameController.text,
-                    _selectedComplaints,
+                if (_categoryNameController.text.isNotEmpty) {
+                  final response = await _firebaseService.addCategory(
+                    _categoryNameController.text,
+                    _selectedSymptoms,
                   );
                   if (response == 200) {
                     ScaffoldMessenger.of(context).showSnackBar(
                       const SnackBar(
-                          content: Text('Symptom added successfully')),
+                        content: Text('Category added successfully')),
                     );
                   } else {
                     ScaffoldMessenger.of(context).showSnackBar(
-                      const SnackBar(content: Text('Failed to add symptom')),
+                      const SnackBar(content: Text('Failed to add category')),
                     );
                   }
                 } else {
@@ -119,8 +119,9 @@ class _SymptomCreationPageState extends State<SymptomCreationPage> {
                   );
                 }
               },
-              child: const Text('Create Symptom', style: TextStyle(fontWeight: FontWeight.bold, fontSize: 18.0)),
+              child: const Text('Create Category', style: TextStyle(fontWeight: FontWeight.bold, fontSize: 18.0)),
             ),
+
             const SizedBox(height: 20),
             ElevatedButton(
               style: const ButtonStyle(
@@ -129,10 +130,10 @@ class _SymptomCreationPageState extends State<SymptomCreationPage> {
                 ),
               onPressed: () {
                 Navigator.of(context).push(
-                  MaterialPageRoute(builder: (context) => const SymptomsListPage()),
+                  MaterialPageRoute(builder: (context) => const CategoriesListPage()),
                 );
               },
-              child: const Text('View All Symptoms', style: TextStyle(fontWeight: FontWeight.bold, fontSize: 18.0)),
+              child: const Text('View All Categories', style: TextStyle(fontWeight: FontWeight.bold, fontSize: 18.0)),
             ),
           ],
         ),
