+++ frontend/s_a_m_e/lib/admin/editDiagnoses.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:s_a_m_e/colors.dart'; 
 import 'package:multi_select_flutter/multi_select_flutter.dart';
-import 'package:s_a_m_e/firebase_service.dart';
-import 'package:s_a_m_e/profilepicture.dart';
+import 'package:s_a_m_e/firebase/firebase_service.dart';
+import 'package:s_a_m_e/account/profilepicture.dart';
 
 class EditDiagnosisPage extends StatelessWidget {
   const EditDiagnosisPage({Key? key}) : super(key: key);
