+++ frontend/s_a_m_e/lib/user/categorieslist.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:s_a_m_e/colors.dart';
-import 'package:s_a_m_e/firebase_service.dart';
-import 'package:s_a_m_e/profilepicture.dart';
+import 'package:s_a_m_e/firebase/firebase_service.dart';
+import 'package:s_a_m_e/account/profilepicture.dart';
 
 class CategoriesListPage extends StatefulWidget {
   const CategoriesListPage({Key? key}) : super(key: key);
