+++ frontend/s_a_m_e/lib/userflow/chooseCategory.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:s_a_m_e/colors.dart';
-import 'package:s_a_m_e/firebase_service.dart';
-import 'package:s_a_m_e/selectSymptoms.dart';
-import 'package:s_a_m_e/user_home.dart';
+import 'package:s_a_m_e/firebase/firebase_service.dart';
+import 'package:s_a_m_e/userflow/selectSymptoms.dart';
+import 'package:s_a_m_e/user/user_home.dart';
 
 class ChooseCategory extends StatefulWidget {
   const ChooseCategory({Key? key}) : super(key: key);
